+++ scripts/main.js
@@ -24,6 +24,7 @@ import { zugaengeHaken, zugaengeEinrichten } from "./zugaenge.js";
 import { anfrageFensterEinrichten } from "./anfrage-fenster.js";
 import { buchFensterEinrichten } from "./ladenbuch.js";
 import { offenenLadenWiederherstellen } from "./vorzeigen.js";
+import { verknuepfungEinrichten, verknuepfungNachtragen } from "./verknuepfung.js";
 import { fensterPassenEinrichten } from "./fensterpassen.js";
 
 function einstellungenEinrichten() {
@@ -117,6 +118,8 @@ Hooks.once("ready", async () => {
   zugaengeEinrichten();
   anfrageFensterEinrichten();
   buchFensterEinrichten();
+  verknuepfungEinrichten();
+  await verknuepfungNachtragen();
   await offenenLadenWiederherstellen();
   await willkommenZeigen();
 });